test(insurance): give planned-change its own test file (T5)
Its coverage lived as an appendix to confirm-org-changes.test.ts, written
against hand-made fixture strings rather than anything the command produces,
which is why the blob preview rendering the same value on both sides went
unnoticed. Move it into test/shared/insurance/planned-change.test.ts, keeping
every existing assertion, and add the cases that were missing: an unresolved-FK
create counting as a create, the counts summing to the row total, the
truncation boundaries at 59/60/61, and a realistic blob rendered through the
preview.
